+++ packages/docs/src/helpers/colors.ts
@@ -10,9 +10,9 @@ import * as tokens from '@synergy-design-system/tokens';
  */
 const getColorAsPalette = (palette: string, useFullTokenName = false) => Object.fromEntries(
   Object.entries(tokens)
-    .filter(([token]) => token.toLowerCase().startsWith(`sdscolor${palette}`))
+    .filter(([token]) => token.toLowerCase().startsWith(`syncolor${palette}`))
     .map(([token, value]) => [
-      useFullTokenName ? token : token.toLowerCase().replace('sdscolor', ''),
+      useFullTokenName ? token : token.toLowerCase().replace('syncolor', ''),
       value,
     ]),
 );